replacing old GridGapControlComponent with GridGapControlComponent2
